pyproject.toml
[project]
name = "gemini-hybrid-chatbot"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "streamlit",
    "google-genai",
    "pdfplumber",
    "faiss-cpu",
    "tavily-python",
    "pandas",
    "numpy",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "reportlab",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["domain", "repository", "service", "component"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
